Patch for the pg truncation adapter to prevent an emtpy list of tables from generating an illegal TRUCATE TABLE statement.

+++ lib/database_cleaner/active_record/truncation.rb
@@ -96,6 +96,7 @@ module ActiveRecord
       end
       
       def truncate_tables(table_names)
+        return if table_names.nil? || table_names.empty?
         execute("TRUNCATE TABLE #{table_names.map{|name| quote_table_name(name)}.join(', ')} #{restart_identity} #{cascade};")
       end
